chore(sandbox): bump to 0.5.7 and dedupe git route helpers
Reuse setup/git and porcelain.parsePorcelainFiles in routes/git.ts;
align sandbox-env chart image tag with the new publish-dialog daemon.
